Make some PKs longs.
Some of these are already 1-2% filled up on our database, and that's only going to go faster. Better to get this out of the way quickly.

+++ SS14.Auth.Shared/Data/AuthHash.cs
@@ -5,7 +5,7 @@ namespace SS14.Auth.Shared.Data;
 
 public class AuthHash
 {
-    public int AuthHashId { get; set; }
+    public long AuthHashId { get; set; }
 
     public Guid SpaceUserId { get; set; }
     public SpaceUser SpaceUser { get; set; }
